git-svn-id: svn+ssh://leetcode.net/home/jreese/backup/srv/svn/repos/rit/3dgames/trunk@1432 ad4848e8-a12d-0410-9a8f-f544a64f585b

+++ GammaDraconis/GammaDraconis/Types/GameObject.cs
@@ -366,8 +366,8 @@ namespace GammaDraconis.Types
         public virtual float getAmmo()
         {
             List<Weapon> weapons = getWeapons(W_TYPE.SECONDARY);
-            int current = 0;
-            int max = 0;
+            float current = 0;
+            float max = 0;
             foreach (Weapon w in weapons)
             {
                 current += w.ammo;
